Project description
This is my second project with Odin Project. Below is the project description.
Don’t forget to commit early & often! You can reference the Commit Message lesson here!
Download the design images and take a look at what you’re going to be creating here. Image One (Full Design), Image Two (Color and Fonts)
The font that’s being used in the images is roboto.
Hero text is the statement that appears at the top of a web page.
There are many ways to tackle a project like this, and it can be overwhelming to look at a blank HTML document and not know where to start. Our suggestion: take it one section at a time. The website you’re creating has 4 main sections (and a footer), so pick one and get it into pretty good shape before moving on. Starting at the top is always a solid plan.
For the section you’re working on, begin by getting all the content onto the page before beginning to style it. In other words, do the HTML and then do the CSS. You’ll probably have to go back to the HTML once you start styling, but bouncing back and forth from the beginning will take more time and may cause more frustration. (Note: you don’t need to use more than one stylesheet. Using only one CSS file is adequate for this project).
Many of the elements on this page are very similar to things you saw in our flexbox exercises… feel free to go back to those if you need a refresher.
Do not worry about making your project look nice on a mobile device. We’ll learn that later.
When you finish, don’t forget to push it up to GitHub!